+++ tests/test_unit/test_response/test_api_redirect_error.py
@@ -4,25 +4,25 @@ import pytest
 from django.core.exceptions import DisallowedRedirect
 from django.utils.http import MAX_URL_REDIRECT_LENGTH
 
-from dmr import APIRedirectError
+from dmr import RedirectTo
 
 
 def test_invalid_redirect_length() -> None:
     """We can't redirect to very long urls."""
     long_url = 'a' * MAX_URL_REDIRECT_LENGTH
     with pytest.raises(DisallowedRedirect, match='Unsafe redirect exceeding'):
-        APIRedirectError(f'custom://{long_url}.com')
+        RedirectTo(f'custom://{long_url}.com')
 
 
 def test_invalid_redirect_scheme() -> None:
     """We can't redirect to untrusted protocols."""
     with pytest.raises(DisallowedRedirect, match='with protocol'):
-        APIRedirectError('custom://url.com')
+        RedirectTo('custom://url.com')
 
 
 def test_invalid_status_code() -> None:
     """We can't redirect with wrong status code."""
     with pytest.raises(DisallowedRedirect, match='3xx statuses'):
-        APIRedirectError('https://url.com', status_code=HTTPStatus.BAD_REQUEST)
+        RedirectTo('https://url.com', status_code=HTTPStatus.BAD_REQUEST)
     with pytest.raises(DisallowedRedirect, match='3xx statuses'):
-        APIRedirectError('https://url.com', status_code=HTTPStatus.IM_USED)
+        RedirectTo('https://url.com', status_code=HTTPStatus.IM_USED)